@@ -1,23 +1,13 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.0</TargetFramework>
-    <UseWindowsForms>true</UseWindowsForms>
-    <RootNamespace>Outlet</RootNamespace>
-    <AssemblyName>Outlet</AssemblyName>
-    <StartupObject>Outlet.Program</StartupObject>
+    <TargetFramework>net5.0</TargetFramework>
     <Nullable>enable</Nullable>
-    <ApplicationIcon />
-    <LangVersion>8.0</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="Properties\AssemblyInfo.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Properties\" />
+    <ProjectReference Include="..\Outlet.ForeignFunctions\Outlet.ForeignFunctions.csproj" />
+    <ProjectReference Include="..\Outlet.StandardLib\Outlet.StandardLib.csproj" />
   </ItemGroup>
 
 </Project>
