@@ -9,8 +9,8 @@ using Decl = Outlet.AST.Declarator;
 namespace Outlet.AST {
 	public class ConstructorDeclaration : FunctionDeclaration {
 
-		public ConstructorDeclaration(Decl decl, List<Decl> argnames, Statement body) 
-            : base(decl, argnames, body) 
+		public ConstructorDeclaration(Decl decl, List<Decl> parameters, List<TypeParameter> typeParameters, Statement body) 
+            : base(decl, parameters, typeParameters, body) 
         {
 		}
 
