@@ -1,7 +1,10 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using Outlet.AST;
 using Outlet.Operands;
+using Outlet.Operators;
+using Outlet.Tokens;
 using Outlet.Types;
 using Type = Outlet.Types.Type;
 
@@ -12,51 +15,44 @@ namespace Outlet.Checking
 
         #region Helpers
 
-        public readonly Stack<bool> DoImpl = new Stack<bool>();
+        //public readonly Stack<bool> DoImpl = new Stack<bool>();
         public readonly Stack<CheckStackFrame> StackFrames = new Stack<CheckStackFrame>();
         public CheckStackFrame CurrentStackFrame => StackFrames.Peek();
-        public class Error : Type
-        {
-            public readonly string Message;
 
-            public Error(string message)
-            {
-                ErrorCount++;
-                if (message != "") Program.ThrowException(message);
-                Message = message;
-            }
-
-            public override bool Is(Type t, out uint level) {
-                level = 0;
-                return false;
-            }
-
-            public override string ToString() => "error";
-        }
-        //private static readonly Type ErrorType = new ProtoClass("error", null, null, null);
-        public static int ErrorCount = 0;
+        private readonly List<Error> CheckingErrors = new List<Error>();
 
         public Checker()
         {
-            StackFrames.Push(CheckStackFrame.Global);
+            StackFrames.Push(CheckStackFrame.Global(Error));
         }
+
+        private void ErrorHandler(Error error)
+        {
+            CheckingErrors.Add(error);
+        }
+
+        public CheckStackFrame GlobalScope => StackFrames.Last();
+
+        public Error Error(string message) => new Error(message, ErrorHandler);
 
         private void Define(Type type, IBindable decl) => CurrentStackFrame.Assign(decl, type);
 
         public void Check(IASTNode program)
         {
-            ErrorCount = 0;
-            if (program is FunctionDeclaration || program is ClassDeclaration)
+            CheckingErrors.Clear();
+            program.Accept(this);
+
+            if (CheckingErrors.Count > 0)
             {
-                DoImpl.Push(false);
-                program.Accept(this);
-                DoImpl.Pop();
-                DoImpl.Push(true);
-                program.Accept(this);
-                DoImpl.Pop();
+                var errorMessage = $"{CheckingErrors.Count} Checking errors encountered\n";
+                foreach(var error in CheckingErrors)
+                {
+                    errorMessage += error.Message + "\n";
+                }
+
+                throw new CheckerException(errorMessage);
             }
-            else program.Accept(this);
-            if (ErrorCount > 0) throw new CheckerException(ErrorCount + " Checking errors encountered");
+
         }
 
         public void EnterScope() => CurrentStackFrame.EnterScope();
@@ -65,7 +61,7 @@ namespace Outlet.Checking
         public CheckStackFrame EnterStackFrame(CheckStackFrame? toPush = null) {
             if (toPush is null)
             {
-                var newFrame = new CheckStackFrame(CurrentStackFrame);
+                var newFrame = new CheckStackFrame(CurrentStackFrame, Error);
                 StackFrames.Push(newFrame);
                 return newFrame;
             }
@@ -76,18 +72,11 @@ namespace Outlet.Checking
 
         public void ExitStackFrame() => StackFrames.Pop();
 
-        //public static Type Error(string message)
-        //{
-        //    ErrorCount++;
-        //    if (message != "") Program.ThrowException(message);
-        //    return ErrorType;
-        //}
-
         public Type Cast(Type from, Type to, string message = "cannot convert type {0} to type {1}")
         {
             if (from is Error) return from;
             if (to is Error) return to;
-            if (!from.Is(to)) return new Error(string.Format(message, from, to));
+            if (!from.Is(to)) return Error(string.Format(message, from, to));
             return to;
         }
 
@@ -95,92 +84,116 @@ namespace Outlet.Checking
 
         #region Declarations
 
+        public void DeclareClass(ClassDeclaration c)
+        {
+            CheckStackFrame statics = new CheckStackFrame(CurrentStackFrame, Error);
+            CheckStackFrame instances = new CheckStackFrame(statics, Error);
+            Class parent = Primitive.Object;
+            if (c.SuperClass != null)
+            {
+                if (c.SuperClass.Accept(this) is MetaType to && to.Stored is Class super)
+                {
+                    parent = super;
+                }
+                else Error("cannot extend anything other than a class");
+            }
+
+            ProtoClass proto = new ProtoClass(c.Name, Error, parent, statics, instances);
+            Define(new MetaType(proto), c.Decl);
+            EnterStackFrame(statics);
+            //foreach (var (id, classConstraint) in c.GenericParameters)
+            //{
+            //    Class constraint = classConstraint?.Accept(this) is TypeObject to && to.Encapsulated is Class co ? co : Primitive.Object;
+            //    CurrentScope.Define(new TypeObject(constraint), id);
+            //}
+
+            foreach (Declaration d in c.StaticDecls) if (d is FunctionDeclaration f) DeclareFunction(f);
+            foreach (var constructor in c.Constructors) DeclareFunction(constructor);
+
+            EnterStackFrame(instances);
+            Define(proto, "this".ToVariable());
+            foreach (Declaration d in c.InstanceDecls) if (d is FunctionDeclaration f) DeclareFunction(f);
+
+            ExitStackFrame();
+            ExitStackFrame();
+        }
+
         public Type Visit(ClassDeclaration c)
         {
-            if (!DoImpl.Peek())
-            {
-                CheckStackFrame statics = new CheckStackFrame(CurrentStackFrame);
-                CheckStackFrame instances = new CheckStackFrame(statics);
-                Class parent = Primitive.Object;
-                if (c.SuperClass != null)
-                {
-                    if (c.SuperClass.Accept(this) is MetaType to && to.Stored is Class super)
-                    {
-                        parent = super;
-                    } else new Error("cannot extend anything other than a class");
-                }
+            //ProtoClass? parent = c.SuperClass != null ? CurrentStackFrame.Get(c.SuperClass) as ProtoClass : null;
 
-                ProtoClass proto = new ProtoClass(c.Name, parent, statics, instances);
-                Define(new MetaType(proto), c.Decl);
-                EnterStackFrame(statics);
-                //foreach (var (id, classConstraint) in c.GenericParameters)
-                //{
-                //    Class constraint = classConstraint?.Accept(this) is TypeObject to && to.Encapsulated is Class co ? co : Primitive.Object;
-                //    CurrentScope.Define(new TypeObject(constraint), id);
-                //}
+            var proto = CurrentStackFrame.Get(c.Decl) is MetaType t && t.Stored is ProtoClass p ? p : throw new CheckerException("Expected protoclass");
+            EnterStackFrame(proto.StaticMembers);
+            foreach (Declaration d in c.StaticDecls) d.Accept(this);
 
-                foreach (Declaration d in c.StaticDecls) if (d is FunctionDeclaration) d.Accept(this);
-                foreach (var constructor in c.Constructors) constructor.Accept(this);
+            EnterStackFrame(proto.InstanceMembers);
+            foreach (Declaration d in c.InstanceDecls) d.Accept(this);
 
-                EnterStackFrame(instances);
-                Define(proto, "this".ToVariable());
-                foreach (Declaration d in c.InstanceDecls) if (d is FunctionDeclaration) d.Accept(this);
+            //if (parent != null) foreach ((string id, Type type) in parent.InstanceMembers) CurrentScope.Define(type, id);
+            foreach (var constructor in c.Constructors) constructor.Accept(this);
+            ExitStackFrame();
+            ExitStackFrame();
 
-                ExitStackFrame();
-                ExitStackFrame();
-            }
-            else
-            {
-                Class? parent;
-                if (c.SuperClass != null) parent = CurrentStackFrame.Get(c.SuperClass) as Class;
-                var proto = CurrentStackFrame.Get(c.Decl) is MetaType t && t.Stored is ProtoClass p ? p : throw new CheckerException("Expected protoclass");
-                EnterStackFrame(proto.StaticMembers);
-                foreach (Declaration d in c.StaticDecls) d.Accept(this);
-
-                EnterStackFrame(proto.InstanceMembers);
-                foreach (Declaration d in c.InstanceDecls) d.Accept(this);
-
-                //if (parent != null) foreach ((string id, Type type) in parent.InstanceMembers) CurrentScope.Define(type, id);
-                foreach (var constructor in c.Constructors) constructor.Accept(this);
-                ExitStackFrame();
-                ExitStackFrame();
-            }
             return Primitive.Void;
         }
 
         public Type Visit(ConstructorDeclaration c) => Visit(c as FunctionDeclaration);
 
-        public Type Visit(FunctionDeclaration f)
+        public void DeclareFunction(FunctionDeclaration f)
         {
-            if (!DoImpl.Peek())
+            if (f.TypeParameters.Any())
+            {
+                FunctionType FillInGenericArgs(List<Type> typeArgs)
+                {
+                    var valid = typeArgs.SameLengthAndAll(f.TypeParameters,
+                        (type, typeParam) => typeParam.Constraint is null || !(Cast(type, typeParam.Constraint.Accept(this)) is Error));
+                    //var types = f.TypeParameters.Zip(typeArgs).ToList().Select((param, type) => null);
+                    return new FunctionType(new (Type, string)[] { (Primitive.Int, "t") }, Primitive.Int);
+                    //throw new NotImplementedException();
+                }
+                var gen = new GenericFunctionType(FillInGenericArgs);
+                Define(gen, f.Decl);
+            }
+            else
             {
                 // Check decl and args first, needed to make function type
-                (Type type, string id)[] args = f.Args.Select(arg =>
-                {
-                    Type curArg = arg.Accept(this);
-                    return (curArg, arg.Identifier);
-                }).ToArray();
+                (Type type, string id)[] parameterTypes = f.Parameters.Select(arg => (arg.Accept(this), arg.Identifier)).ToArray();
                 Type returnType = f.Decl.Accept(this) is Type t ? t : throw new CheckerException("Expected Type");
-                FunctionType ft = new FunctionType(args, returnType);
+                FunctionType ft = new FunctionType(parameterTypes, returnType);
                 // define the header using the function type from above
                 Define(ft, f.Decl);
-                return ft;
             }
-            else if (CurrentStackFrame.Resolve(f.Decl, out Type type, out uint _, out uint _))
+        }
+
+        public Type Visit(FunctionDeclaration f)
+        {
+            if (CurrentStackFrame.Resolve(f.Decl, out Type type, out uint _, out uint _))
             {
-                FunctionType ft = type is FunctionType fnt ? fnt : throw new CheckerException("Expected Function type");
-                // enter the function scope and define the args;
+                // enter the function scope and define the args and type args
                 EnterStackFrame();
-                ft.Args.Zip(f.Args).ToList().ForEach(arg => Define(arg.First.type, arg.Second));
+                if(f.TypeParameters.Any() && type is GenericFunctionType generic)
+                {
+                    foreach (var typeParam in f.TypeParameters)
+                    {
+                        if ((typeParam.Constraint?.Accept(this) ?? new MetaType(Primitive.Object)) is MetaType constraintType)
+                        {
+                            Define(constraintType, typeParam);
+                        }
+                        else Error($"Generic parameter {typeParam.Identifier} constraint was not a valid type");
+                    }
+                }
+                FunctionType ft = type is FunctionType fnt ? fnt : throw new CheckerException("Expected Function type");
+
+                foreach (var ((argType, _), argDecl) in ft.Parameters.Zip(f.Parameters)) Define(argType, argDecl);
                 // check the body now that its header and args have been defined
                 Type body = f.Body.Accept(this);
                 if (f is ConstructorDeclaration)
                 {
-                    if (body != Primitive.Void) return new Error("constructor cannot return value");
+                    if (body != Primitive.Void) return Error("constructor cannot return value");
                 }
-                else if (body == null || body == Primitive.Void)
+                else if (body == Primitive.Void)
                 {
-                    if (ft.ReturnType != Primitive.Void) return new Error("function " + f.Decl.Identifier + "not all code paths return a value");
+                    if (ft.ReturnType != Primitive.Void) return Error("function " + f.Decl.Identifier + "not all code paths return a value");
                 }
                 else Cast(body, ft.ReturnType, f.Decl.Identifier + " function definition invalid, expected {1}, returned {0}");
                 f.LocalCount = CurrentStackFrame.Count;
@@ -190,21 +203,50 @@ namespace Outlet.Checking
             else throw new UnexpectedException("Could not resolve function type");
         }
 
+        public Type Visit(OperatorOverloadDeclaration o)
+        {
+            throw new NotImplementedException("In development");
+            //if (!DoImpl.Peek())
+            //{
+            //    switch (o.Operator)
+            //    {
+            //        case BinaryOperator b:
+            //            if (o.Parameters.Count != 2) return Error($"Binary operator {b.Name} requires two parameters to overload");
+            //            break;
+            //        case UnaryOperator u:
+            //            if (o.Parameters.Count != 1) return Error($"Unary operator {u.Name} requires one parameter to overload");
+            //            break;
+            //        default:
+            //            throw new UnexpectedException("No other operator types");
+            //    }
+            //    var res = Visit(o as FunctionDeclaration);
+            //    return res;
+            //}
+            //else
+            //{
+            //    return Visit(o as FunctionDeclaration);
+            //}
+        }
+
         public Type Visit(VariableDeclaration v)
         {
             Type? init = v.Initializer?.Accept(this);
             if (v.Decl.IsVar)
             {
-                if (init is null) return new Error($"Cannot determine type of var {v.Name} without initializer");
+                if (init is null) return Error($"Cannot determine type of var {v.Name} without initializer");
                 if (init is MetaType meta) Define(meta, v.Decl);
                 else Define(init, v.Decl);
             }
             else
             {
                 Type decl = v.Decl.Accept(this);
-                if (init != null) Cast(init, decl);
-                if (init is MetaType meta) Define(meta, v.Decl);
-                else Define(decl, v.Decl);
+                if (init != null)
+                {
+                    var casted = Cast(init, decl);
+                    if (casted is Error e) return e;
+                    if (casted is MetaType meta) Define(meta, v.Decl);
+                    else Define(casted, v.Decl);
+                }
             }
             return Primitive.Void;
         }
@@ -216,21 +258,27 @@ namespace Outlet.Checking
         public Type Visit(Access a)
         {
             Type elem = a.Collection.Accept(this);
-            Type? idxType = a.Index.Length > 0 ? a.Index[0].Accept(this) : null;
-            if (elem is MetaType meta && meta.Stored is Class c)
+            var idxTypes = a.Index.Select(idx => idx.Accept(this));
+            // array types are defined with empty braces []
+            if (elem is MetaType m && !idxTypes.Any()) return new MetaType(new ArrayType(m.Stored));
+            if (elem is IGenericType gt && idxTypes.All(type => type is MetaType))
             {
-                if (idxType is MetaType)
-                    return new Error("Generics not supported yet");
-                // array types are defined with empty braces []
-                if (idxType == null)
-                    return new MetaType(new ArrayType(meta.Stored));
+                var unwrappedTypes = idxTypes.OfType<MetaType>().Select(m => m.Stored);
+                return gt.WithTypeArguments(unwrappedTypes);
             }
-            if (a.Index.Length != 1)
-                return new Error("array access requires exactly 1 index");
-            if (idxType != Primitive.Int)
-                return new Error("only ints can be used to index into an array, found: " + idxType?.ToString());
-            if (elem is ArrayType at) return at.ElementType;
-            return new Error("type " + elem.ToString() + " is not accessable by array access operator []");
+            //if (elem is GenericClass gt && idxTypes.All(type => type is MetaType))
+            //{
+            //    // TODO same here generate ProtoClass based on type parameters
+            //    return Error("Generics not supported yet");
+            //}
+            if (elem is ArrayType at)
+            {
+                if (a.Index.Length != 1)
+                    return Error("array access requires exactly 1 index");
+                if (idxTypes.Count() != 1 || idxTypes.First() != Primitive.Int) return Error("only ints can be used to index into an array, found: " + idxTypes?.ToList().ToListString() ?? "empty");
+                else return at.ElementType;
+            }
+            return Error("type " + elem.ToString() + " is not accessable by array access operator []");
         }
 
         public Type Visit(As a)
@@ -238,12 +286,12 @@ namespace Outlet.Checking
             a.Left.Accept(this);
             Type r = a.Right.Accept(this);
             if (r is MetaType castedTo) return castedTo.Stored;
-            return new Error("the right side of an is expression must be a type, found: " + r.ToString());
+            return Error("the right side of an is expression must be a type, found: " + r.ToString());
         }
 
         public Type Visit(Assign a) => a.Left switch
         {
-            MemberAccess d when d.ArrayLength => new Error("cannot assign to an array length"),
+            MemberAccess d when d.ArrayLength => Error("cannot assign to an array length"),
             Expression left when left is Variable || left is MemberAccess =>
                 (a.Left.Accept(this), a.Right.Accept(this)) switch
                 {
@@ -251,7 +299,7 @@ namespace Outlet.Checking
                     (_, Error e) => e,
                     (Type l, Type r) => Cast(r, l)
                 },
-            _ => new Error("illegal assignment, can only assign to variables and fields")
+            _ => Error("illegal assignment, can only assign to variables and fields")
         };
 
         public Type Visit(Binary b) => (b.Left.Accept(this), b.Right.Accept(this)) switch
@@ -259,18 +307,19 @@ namespace Outlet.Checking
             (Error e, _) => e,
             (_, Error e) => e,
             (Type left, Type right) => 
-                b.Overloads.FindBestMatch(left, right) switch
-                {
-                    null => new Error($"binary operator not defined for {left} {b.Op} {right}"),
-                    BinOp op => (b.Oper = op).GetResultType()
-                }
+                // Try to find built in operator first
+                (b.Oper = b.Overloads.FindBestMatch(left, right))?.GetResultType() ??
+                (GlobalScope.Has(b.Op) ? 
+                    // If it's user defined it can be found in global scope
+                    GlobalScope.Get(b.Op.ToVariable()) : 
+                    Error($"binary operator not defined for {left} {b.Op} {right}"))
         };
 
         public Type Visit(Call c)
         {
             Type calltype = c.Caller.Accept(this);
             if (calltype is Error) return calltype;
-            if (calltype is MetaType t && t.Stored is ProtoClass proto)
+            if (calltype is MetaType t && t.Stored is ProtoClass)
             {
                 // Rather than have to store the overload id within the call it is easier to turn a constructor call into
                 // a dereference to static field "" where the constructor lives
@@ -281,43 +330,39 @@ namespace Outlet.Checking
             Type[] argtypes = c.Args.Select(x => x.Accept(this)).ToArray();
             if (calltype is FunctionType functype)
             {
-                var funcargs = functype.Args.ToArray();
-                bool argsMatch = funcargs.SameLengthAndAll(argtypes, (arg, argType) => !(Cast(argType, arg.type, "Could not cast {0} to {1}") is Error));
-                return argsMatch ? functype.ReturnType :
-                    new Error(c.Caller + " expects " + "(" + funcargs.Select(x => x.type).ToList().ToListString() + ") found: (" + argtypes.ToList().ToListString() + ")");
+                return functype.Valid(out uint _, argtypes) ?
+                    functype.ReturnType :
+                    Error($"{c.Caller} expects ({functype.Parameters.Select(x => x.type).ToList().ToListString()}) found: ({argtypes.ToList().ToListString()})");
             }
             if (calltype is MethodGroupType mgt)
             {
                 (FunctionType? bestMatch, uint? id) = mgt.FindBestMatch(argtypes);
-                if(bestMatch is null || id is null) return new Error("No overload could be found for (" + argtypes.ToList().ToListString() + ")");
-                if (c.Caller is Variable v && v.ResolveLevel.HasValue)
+                if(bestMatch is null || id is null) return Error("No overload could be found for (" + argtypes.ToList().ToListString() + ")");
+                IBindable caller = c.Caller is Variable v ? v : c.Caller is MemberAccess ma ? ma.Member : throw new UnexpectedException("Caller is not able to be overloaded");
+                if(caller.ResolveLevel.HasValue)
                 {
-                    v.Bind(id.Value, v.ResolveLevel.Value);
+                    caller.Bind(id.Value, caller.ResolveLevel.Value);
                 }
-                else if (c.Caller is MemberAccess ma && ma.Member.ResolveLevel.HasValue)
-                {
-                    ma.Member.Bind(id.Value, ma.Member.ResolveLevel.Value);
-                }
-                else throw new UnexpectedException("Caller is not able to be overloaded");
+                else throw new UnexpectedException("Unresolved caller");
                 return bestMatch.ReturnType;
             }
-            return new Error("type " + calltype + " is not callable");
+            return Error($"type {calltype} is not callable");
         }
 
         public Type Visit(Declarator d) => d.Type?.Accept(this) switch
         {
             MetaType meta => meta.Stored,
-            Primitive t when t == Primitive.MetaType => new Error("Declared type must be a check time constant"),
-            Type invalid => new Error($"Declaration requires valid type, found: {invalid}"),
-            null => new Error($"Cannot use var here")
+            Primitive t when t == Primitive.MetaType => Error("Declared type must be a check time constant"),
+            Type invalid => Error($"Declaration requires valid type, found: {invalid}"),
+            null => Error($"Cannot use var here")
         };
 
         public Type Visit(TupleAccess t) => t.Left.Accept(this) switch
         {
             Error e => e,
             TupleType tt when tt.Types.Length > t.Member => tt.Types[t.Member],
-            TupleType tt => new Error($"cannot access element {t.Member} of tuple type {tt} which has {tt.Types.Length} elements"),
-            Type type => new Error($"Cannot reference member {t.Member} of non tuple type {type}")
+            TupleType tt => Error($"cannot access element {t.Member} of tuple type {tt} which has {tt.Types.Length} elements"),
+            Type type => Error($"Cannot reference member {t.Member} of non tuple type {type}")
         };
 
         public Type Visit(MemberAccess d) => d.Left.Accept(this) switch
@@ -326,7 +371,7 @@ namespace Outlet.Checking
             ArrayType _ when d.ArrayLengthAccess() => Primitive.Int,
             ProtoClass instances => instances.GetInstanceMemberType(d.Member),
             MetaType t when t.Stored is ProtoClass statics => statics.GetStaticMemberType(d.Member),
-            Type other => new Error($"cannot dereference type: {other}")
+            Type other => Error($"cannot dereference type: {other}")
         };
 
         public Type Visit(Is i)
@@ -334,13 +379,13 @@ namespace Outlet.Checking
             i.Left.Accept(this);
             Type r = i.Right.Accept(this);
             if (r is MetaType || r == Primitive.MetaType) return Primitive.Bool;
-            return new Error("the right side of an is expression must be a type, found: " + r.ToString());
+            return Error("the right side of an is expression must be a type, found: " + r.ToString());
         }
 
         public Type Visit(Lambda l) => (l.Left.Accept(this), l.Right.Accept(this)) switch
         {
-            (MetaType args, MetaType result) => new Error("NOT IMPLEMENTED"),  //new TypeObject(new FunctionType())
-            _ => new Error("Lambdas currently old work for types")
+            (MetaType args, MetaType result) => Error("NOT IMPLEMENTED"),  //new TypeObject(new FunctionType())
+            _ => Error("Lambdas currently old work for types")
         };
 
         public Type Visit(ListLiteral l)
@@ -350,7 +395,7 @@ namespace Outlet.Checking
 
         public Type Visit<E>(Literal<E> c) where E : struct => c.Type;
 
-        public Type Visit(StringLiteral s) => Primitive.String;
+        public Type Visit(AST.StringLiteral s) => Primitive.String;
 
         public Type Visit(NullExpr n) => Primitive.Object;
 
@@ -364,7 +409,7 @@ namespace Outlet.Checking
         public Type Visit(Ternary t) => (t.Condition.Accept(this), t.IfTrue.Accept(this), t.IfFalse.Accept(this)) switch
         {
             (Primitive condition, _, _) when condition != Primitive.Bool => 
-                new Error($"Ternary condition requires a boolean, found a {condition}"),
+                Error($"Ternary condition requires a boolean, found a {condition}"),
             (_, Error e, _) => e,
             (_, _, Error e) => e,
             (_, Type iftrue, Type iffalse) => Type.CommonAncestor(iftrue, iffalse)
@@ -372,7 +417,7 @@ namespace Outlet.Checking
 
         public Type Visit(TupleLiteral t)
         {
-            if (t.Args.Length == 1) return t.Args[0].Accept(this);
+            if (t.Args.Length == 1) return t.Args.First().Accept(this);
             var types = t.Args.Select(arg => arg.Accept(this)).ToArray();
             if (types.All(type => type is MetaType)) return new MetaType(new TupleType(types));
             else return new TupleType(types);
@@ -383,7 +428,7 @@ namespace Outlet.Checking
             Type input = u.Expr.Accept(this);
             if (input is Error) return input;
             var op = u.Overloads.FindBestMatch(input);
-            if (op == null) return new Error("unary operator " + u.Op + " is not defined for type " + input.ToString());
+            if (op == null) return Error("unary operator " + u.Op + " is not defined for type " + input.ToString());
             u.Oper = op;
             return op.GetResultType();
         }
@@ -393,9 +438,8 @@ namespace Outlet.Checking
             if(CurrentStackFrame.Resolve(v, out Type type, out uint level, out uint id))
             {
                 v.Bind(id, level);
-                return type;
             }
-            return new Error("variable " + v.Identifier + " could not be resolved");
+            return type;
         }
 
         #endregion
@@ -404,32 +448,26 @@ namespace Outlet.Checking
 
         public Type Visit(Block b)
         {
-            EnterScope();
-            DoImpl.Push(false);
+            if(!b.IsProgram) EnterScope();
             // Forward Declaration of Classes
-            foreach (ClassDeclaration cd in b.Classes)
-            {
-                cd.Accept(this);
-            }
+            foreach (ClassDeclaration cd in b.Classes) DeclareClass(cd);
             // Forward Declaration of Functions
-            foreach (FunctionDeclaration fd in b.Functions)
+            foreach (FunctionDeclaration func in b.Functions) DeclareFunction(func);
+            foreach (var overload in b.OverloadedOperators)
             {
-                fd.Accept(this);
+                overload.Accept(this);
             }
-            DoImpl.Pop();
-            DoImpl.Push(true);
             Type? ret = null;
             foreach (IASTNode d in b.Lines)
             {
                 Type temp = d.Accept(this);
-                if (ret != null) return new Error("unreachable code detected");
-                if (d is Statement && !(d is Expression) && temp != null)
+                if (ret != null) return Error("unreachable code detected");
+                if (d is Statement and not Expression && temp != null)
                 {
                     ret = temp;
                 }
             }
-            DoImpl.Pop();
-            ExitScope();
+            if(!b.IsProgram) ExitScope();
             return ret is null ? Primitive.Void : ret;
         }
 
@@ -444,10 +482,10 @@ namespace Outlet.Checking
                 Define(loopvar, f.LoopVar);
                 Type body = f.Body.Accept(this);
                 ExitScope();
-                if (f.Body is Statement && !(f.Body is Expression) && body != null) return body;
+                if (f.Body is Statement and not Expression && body != null) return body;
                 return Primitive.Void;
             }
-            return new Error("only array types are iterable, found:" + collection.ToString());
+            return Error("only array types are iterable, found:" + collection.ToString());
         }
 
         public Type Visit(IfStatement i)
@@ -456,9 +494,9 @@ namespace Outlet.Checking
             Cast(cond, Primitive.Bool, "if statement condition requires a boolean, found a {0}");
             Type iftrue = i.Iftrue.Accept(this);
             Type? iffalse = i.Iffalse?.Accept(this);
-            if (i.Iftrue is Statement && !(i.Iftrue is Expression) && iftrue != null)
+            if (i.Iftrue is Statement and not Expression && iftrue != null)
             {
-                if (i.Iffalse is Statement && !(i.Iffalse is Expression) && iffalse != null)
+                if (i.Iffalse is Statement and not Expression && iffalse != null)
                 {
                     // TODO should return common ancestor
                     return iftrue;
@@ -477,7 +515,7 @@ namespace Outlet.Checking
             Type cond = w.Condition.Accept(this);
             Cast(cond, Primitive.Bool, "while loop condition requires a boolean, found a {0}");
             Type body = w.Body.Accept(this);
-            if (w.Body is Statement && !(w.Body is Expression) && body != null) return body;
+            if (w.Body is Statement and not Expression && body != null) return body;
             return Primitive.Void;
         }
 
@@ -488,7 +526,7 @@ namespace Outlet.Checking
                 foreach (var (id, type) in staticClass.GetStaticMemberTypes())
                 {
                     // TODO restore functionality
-                    //CurrentScope.Define(type, id);
+                    CurrentStackFrame.Assign(new Variable(id), type);
                 }
             {
                 return Primitive.Void;
