@@ -7,7 +7,7 @@ namespace Outlet.Interpreting
     public class StackFrame : IStackFrame<Operand>
     {
         public (string Id, Operand Value)[] LocalVariables { get; private set; }
-        private readonly IStackFrame<Operand>? Parent;
+        private readonly StackFrame? Parent;
         public string Call { get; private set; }
 
         public static readonly StackFrame Global = new StackFrame();
@@ -16,21 +16,23 @@ namespace Outlet.Interpreting
         {
             Call = "Global scope";
             int index = 0;
-            LocalVariables = new (string, Operand)[ForeignFunctions.NativeTypes.Count];
-            foreach(string type in ForeignFunctions.NativeTypes.Keys)
+            LocalVariables = new (string, Operand)[NativeOutletTypes.NativeTypes.Count];
+            foreach(string type in NativeOutletTypes.NativeTypes.Keys)
             {
-                LocalVariables[index++] = (type, new TypeObject(ForeignFunctions.NativeTypes[type]));
+                LocalVariables[index++] = (type, new TypeObject(NativeOutletTypes.NativeTypes[type]));
             }
         }
 
-        public StackFrame(IStackFrame<Operand> parent, uint localCount, string call)
+        public StackFrame(StackFrame parent, uint localCount, string call)
         {
             Parent = parent;
             LocalVariables = new (string, Operand)[localCount];
             Call = call;
         }
 
-        public Operand Get(IBindable variable, uint level = 0)
+        public Operand Get(IBindable variable) => Get(variable, level: 0);
+
+        private Operand Get(IBindable variable, uint level = 0)
         {
             if (variable.ResolveLevel > level)
             {
@@ -38,10 +40,12 @@ namespace Outlet.Interpreting
                 else return Parent.Get(variable, level + 1);
             }
             else if(variable.LocalId.HasValue) return LocalVariables[variable.LocalId.Value].Value;
-            else throw new UnexpectedException("Variable was not resolved");
+            else throw new UnexpectedException($"Variable {variable.Identifier} was not resolved");
         }
 
-        public void Assign(IBindable variable, Operand value, uint level = 0)
+        public void Assign(IBindable variable, Operand value) => Assign(variable, value, level: 0);
+
+        private void Assign(IBindable variable, Operand value, uint level = 0)
         {
             if (this == Global) 
             {
@@ -55,7 +59,7 @@ namespace Outlet.Interpreting
                 else Parent.Assign(variable, value, level + 1);
             }
             else if(variable.LocalId.HasValue) LocalVariables[variable.LocalId.Value] = (variable.Identifier, value);
-            else throw new UnexpectedException("Variable was not resolved");
+            else throw new UnexpectedException($"Variable {variable.Identifier} was not resolved");
         }
 
         public IEnumerable<(string Id, Operand Value)> List() => LocalVariables;
