@@ -7,194 +7,177 @@ using Outlet.Tokens;
 using Outlet.AST;
 
 namespace Outlet.Parsing {
-	public static partial class Parser {
+	public partial class Parser {
 
-		public static IASTNode NextDeclaration(LinkedList<Token> tokens) {
+		private IASTNode NextDeclaration() {
 			#region helper
-			bool Match(Token s) {
-				if(tokens.Count > 0 && s.Equals(tokens.First())) {
-					tokens.RemoveFirst(); return true;
-				} else return false;
+			List<TypeParameter> ParseGenericParameters()
+            {
+				List<TypeParameter> parameters = new List<TypeParameter>();
+				if (Match(DelimeterToken.LeftBrace))
+				{
+					do
+					{
+						string genericId = ConsumeType<Identifier>(expected: "generic parameter identifier").Name;
+						Expression? genericConstraint = Match(Keyword.Extends) ? NextExpression() : null;
+						parameters.Add(new TypeParameter(genericConstraint, genericId));
+					} while (Match(DelimeterToken.Comma));
+					Consume(DelimeterToken.RightBrace, expected: "] to close generic type definition");
+				}
+				return parameters;
 			}
-			void Consume(Token s, string error) {
-				if(tokens.Count == 0 || tokens.Dequeue() != s) throw new OutletException("Syntax Error: " + error);
-			}
-			T ConsumeType<T>(string error) where T : Token {
-				if(tokens.Count > 0 && tokens.Dequeue() is T t) return t;
-				else throw new OutletException("Syntax Error: " + error);
-			}
-			(List<Declarator>, Statement) ProtoType() {
+			(List<Declarator>, List<TypeParameter>, Statement) ProtoType() {
+				var genericParameters = ParseGenericParameters();
+				Consume(DelimeterToken.LeftParen, expected: "( after generic parameters");
 				List<Declarator> argnames = new List<Declarator>();
-				while(tokens.Count > 0 && tokens.First() != Delimeter.RightParen) {
+				while(PeekNextTokenExistsAndIsnt(DelimeterToken.RightParen)) {
 					do {
-						if(NextStatement(tokens) is Declarator paramdecl) {
+						if(NextStatement() is Declarator paramdecl) {
 							argnames.Add(paramdecl);
 						} else throw new OutletException("function parameters expected in type id format");
-					} while(Match(Delimeter.Comma));
+					} while(Match(DelimeterToken.Comma));
 				}
-				Consume(Delimeter.RightParen, " expected ) after function args");
+				Consume(DelimeterToken.RightParen, expected: ") after function args");
 				Statement body;
-				if(Match(Operator.Lambda)) {
-					body = NextExpression(tokens);
-					if(tokens.Count != 0)
-						Consume(Delimeter.SemiC, "expected ; after inline function");
-				} else body = NextStatement(tokens);
-				return (argnames, body);
+				if(Match(OperatorToken.Lambda)) {
+					body = NextExpression();
+					if(Tokens.Count != 0)
+						Consume(DelimeterToken.SemiC, expected: "; after inline function");
+				} else body = NextStatement();
+				return (argnames, genericParameters, body);
 			}
 			#endregion
 			VariableDeclaration VarDeclaration(Declarator decl) {
 				Expression? initializer = null;
-				if(Match(Operator.Equal)) initializer = NextExpression(tokens);
-				if(tokens.Count != 0)
-					Consume(Delimeter.SemiC, "expected ; after declaring a variable");
+				if(Match(OperatorToken.Equal)) initializer = NextExpression();
+				if(Tokens.Count != 0)
+					Consume(DelimeterToken.SemiC, expected: "; after declaring a variable");
 				return new VariableDeclaration(decl, initializer);
 			}
 			FunctionDeclaration FunctionDef(Declarator decl) {
-				(List<Declarator> argnames, Statement body) = ProtoType();
-				return new FunctionDeclaration(decl, argnames, body);
+				(List<Declarator> parameters, List<TypeParameter> typeParams, Statement body) = ProtoType();
+				return new FunctionDeclaration(decl, parameters, typeParams, body);
 			}
 			ConstructorDeclaration ConstructDef(Declarator decl) {
-				(List<Declarator> argnames, Statement body) = ProtoType();
-				return new ConstructorDeclaration(decl, argnames, body);
+				(List<Declarator> parameters, List<TypeParameter> typeParams, Statement body) = ProtoType();
+				return new ConstructorDeclaration(decl, parameters, typeParams, body);
 			}
+			OperatorOverloadDeclaration OperatorOverloadDef(Declarator decl, OperatorToken op)
+			{
+				(List<Declarator> parameters, List<TypeParameter> typeParams, Statement body) = ProtoType();
+				return new OperatorOverloadDeclaration(decl, op, parameters, typeParams, body);
+            }
 			ClassDeclaration ClassDef() {
-                List<(string id, Variable? constraint)> genericParameters = new List<(string, Variable?)>();
                 List<Declaration> instance = new List<Declaration>();
 				List<Declaration> statics = new List<Declaration>();
 				List<ConstructorDeclaration> constructors = new List<ConstructorDeclaration>();
 				Variable? superclass = null;
-                Identifier name = ConsumeType<Identifier>("Expected class identifier");
-                if (Match(Delimeter.LeftBrace))
-                {
-                    string genericId = ConsumeType<Identifier>("Generic class must have at least one identifier as a generic parameter").Name;
-                    if (Match(Keyword.Extends)) genericParameters.Add((genericId, 
-                        new Variable(ConsumeType<Identifier>("expected class constraint on generic parameter " + genericId).Name)));
-                    else genericParameters.Add((genericId, null));
-                    while(tokens.Count > 0 && tokens.First() != Delimeter.RightBrace)
-                    {
-                        Consume(Delimeter.Comma, "commas must be used between generic parameters");
-                        genericId = ConsumeType<Identifier>("Generic class parameters must be identifiers").Name;
-
-                        if (Match(Keyword.Extends)) genericParameters.Add((genericId,
-                        new Variable(ConsumeType<Identifier>("expected class constraint on generic parameter " + genericId).Name)));
-                        else genericParameters.Add((genericId, null));
-                    }
-                    Consume(Delimeter.RightBrace, "expected ] to close generic type definition");
-                }
+                Lexeme nameLexeme = ConsumeTypeGetLexeme<Identifier>(expected: "class identifier");
+				Identifier name = nameLexeme.InnerToken as Identifier;
+				var genericParameters = ParseGenericParameters();
                 if (Match(Keyword.Extends)) {
-					 superclass = new Variable(ConsumeType<Identifier>("expected name of super class after extends keyword").Name);
+					 superclass = new Variable(ConsumeType<Identifier>(expected: "name of super class after extends keyword").Name);
 				}
-				if(Match(Delimeter.LeftCurly)) {
+				if(Match(DelimeterToken.LeftCurly)) {
 					while(true) {
-						if(Match(Delimeter.RightCurly)) break;
-						if(tokens.Count == 0) throw new OutletException("expected } after class definition");
+						if(Match(DelimeterToken.RightCurly)) break;
+						if(Tokens.Count == 0) throw new OutletException("expected } after class definition");
 						if(Match(name)) {
-							if(Match(Delimeter.LeftParen)) {
+							if(PeekMatch(DelimeterToken.LeftParen) || PeekMatch(DelimeterToken.LeftBrace)) {
 								Declarator constr = new Declarator(new Variable(name.Name), "");
 								constructors.Add(ConstructDef(constr));
 								continue;
-							} else tokens.AddFirst(name);
+							} else Tokens.AddFirst(nameLexeme);
 						}
 						bool isstatic = Match(Keyword.Static);
-						Statement nextfield = NextStatement(tokens);
+						Statement nextfield = NextStatement();
 						if(nextfield is Declarator df) {
-							Declaration curdecl = Match(Delimeter.LeftParen) ? FunctionDef(df) as Declaration : VarDeclaration(df);
+							Declaration curdecl = PeekMatch(DelimeterToken.LeftParen) || PeekMatch(DelimeterToken.LeftBrace) ? FunctionDef(df) as Declaration : VarDeclaration(df);
 							(isstatic ? statics : instance).Add(curdecl);
 						} else throw new OutletException("statement: " + nextfield.ToString() + " must be inside a function body");
 					}
 				}
-				if(constructors.Count == 0) constructors.Add(new ConstructorDeclaration(new Declarator(new Variable(name.Name), ""), new List<Declarator>(), new Block(new List<IASTNode>(), new List<FunctionDeclaration>(), new List<ClassDeclaration>())));
+				if(constructors.Count == 0) constructors.Add(new ConstructorDeclaration(new Declarator(new Variable(name.Name), ""), new List<Declarator>(), new List<TypeParameter>(), Block.Empty()));
 				return new ClassDeclaration(name.Name, superclass, genericParameters, constructors, instance, statics);
 			}
 			if(Match(Keyword.Class)) return ClassDef();
-			Statement next = NextStatement(tokens);
-
+			Statement next = NextStatement();
 			if(next is Declarator d) {
-				if(Match(Delimeter.LeftParen)) return FunctionDef(d);
+				if(d.IsOperatorOverload)
+                {
+					var op = ConsumeType<OperatorToken>(expected: "operator following overload");
+					// TODO this should probably be a PeekMatch and also check for LeftBrace for generic case
+					Consume(DelimeterToken.LeftParen, expected: "( before operator overload args");
+					return OperatorOverloadDef(new Declarator(d.Type, op.ToString()), op);
+                }
+				if(PeekMatch(DelimeterToken.LeftParen) || PeekMatch(DelimeterToken.LeftBrace)) return FunctionDef(d);
 				else return VarDeclaration(d);
 			}
 			return next;
 		}
 
-        public static Statement NextStatement(LinkedList<Token> tokens) {
-            #region helper
-            bool Match(Token s) {
-                if (tokens.Count > 0 && tokens.First() == s) {
-                    tokens.Dequeue(); return true;
-                } else return false;
-            }
-            void Consume(Token s, string error) {
-                if (tokens.Count == 0 || tokens.Dequeue() != s) throw new OutletException("Syntax Error: " + error);
-            }
-            T ConsumeType<T>(string error) where T : Token {
-                if (tokens.Count > 0 && tokens.Dequeue() is T t) return t;
-                else throw new OutletException("Syntax Error: " + error);
-            }
-            #endregion
+        private Statement NextStatement() {
             Statement Scope() {
-                List<IASTNode> lines = new List<IASTNode>();
-                List<FunctionDeclaration> funcs = new List<FunctionDeclaration>();
-                List<ClassDeclaration> classes = new List<ClassDeclaration>();
-                while (tokens.Count > 0 && tokens.First() != Delimeter.RightCurly) {
-                    var nextdecl = NextDeclaration(tokens);
-                    if (nextdecl is FunctionDeclaration fd) funcs.Add(fd);
-                    if (nextdecl is ClassDeclaration cd) {
-                        if (cd.SuperClass == null) classes.Insert(0, cd);
-                        else classes.Add(cd);
-                    }
-                    lines.Add(nextdecl);
-                }
-                Consume(Delimeter.RightCurly, "Expected } to close code block");
-                return new Block(lines, funcs, classes);
+				var block = ParseBlock();
+				Consume(DelimeterToken.RightCurly, expected: "} to close code block");
+                return block;
             }
             Statement IfStatement() {
-                Consume(Delimeter.LeftParen, "Expected ( after if");
-                Expression condition = NextExpression(tokens);
-                Consume(Delimeter.RightParen, "Expected ) after if condition");
-                Statement iftrue = NextStatement(tokens);
-                Statement? ifelse = Match(Keyword.Else) ? NextStatement(tokens) : null;
+                Consume(DelimeterToken.LeftParen, expected: "( after if");
+                Expression condition = NextExpression();
+                Consume(DelimeterToken.RightParen, expected: ") after if condition");
+                Statement iftrue = NextStatement();
+                Statement? ifelse = Match(Keyword.Else) ? NextStatement() : null;
                 return new IfStatement(condition, iftrue, ifelse);
             }
             Statement WhileLoop() {
-                Consume(Delimeter.LeftParen, "Expected ( after while");
-                Expression condition = NextExpression(tokens);
-                Consume(Delimeter.RightParen, "Expected ) after while condition");
-                Statement iftrue = NextStatement(tokens);
+                Consume(DelimeterToken.LeftParen, expected: "( after while");
+                Expression condition = NextExpression();
+                Consume(DelimeterToken.RightParen, expected: ") after while condition");
+                Statement iftrue = NextStatement();
                 return new WhileLoop(condition, iftrue);
             }
             Statement ForLoop() {
-                Consume(Delimeter.LeftParen, "Expected ( after for");
-                Statement s = NextStatement(tokens);
+                Consume(DelimeterToken.LeftParen, expected: "( after for");
+                Statement s = NextStatement();
                 if (s is Declarator loopvar) {
-                    Consume(Keyword.In, "expected in after for loop variable");
-                    Expression collection = NextExpression(tokens);
-                    Consume(Delimeter.RightParen, "Expected ) after for loop collection");
-                    Statement body = NextStatement(tokens);
+                    Consume(Keyword.In, expected: "'in' after for loop variable");
+                    Expression collection = NextExpression();
+                    Consume(DelimeterToken.RightParen, expected: ") after for loop collection");
+                    Statement body = NextStatement();
                     return new ForLoop(loopvar, collection, body);
                 }
-                throw new OutletException("expected type followed by an identifier to use as a loop variable");
+				throw new OutletException("expected type followed by an identifier to use as a loop variable");
             }
             Statement Return() {
-                Expression retexpr = NextExpression(tokens);
-                Consume(Delimeter.SemiC, "expected ; after return statement");
+                Expression retexpr = NextExpression();
+                Consume(DelimeterToken.SemiC, expected: "; after return statement");
                 return new ReturnStatement(retexpr);
             }
             Statement Using()
             {
-                Expression used = NextExpression(tokens);
-                Consume(Delimeter.SemiC, "expected ; after using statement");
+                Expression used = NextExpression();
+                Consume(DelimeterToken.SemiC, expected: "; after using statement");
                 return new UsingStatement(used);
             }
-			if(Match(Delimeter.LeftCurly)) return Scope();
+			if(Match(DelimeterToken.LeftCurly)) return Scope();
 			if(Match(Keyword.If)) return IfStatement();
 			if(Match(Keyword.For)) return ForLoop();
 			if(Match(Keyword.While)) return WhileLoop();
 			if(Match(Keyword.Return)) return Return();
             if(Match(Keyword.Using)) return Using();
-			Expression? expr = Match(Keyword.Var) ? null : NextExpression(tokens);
-			if((Match(Delimeter.SemiC) || tokens.Count == 0) && expr != null) return expr;
-			Identifier id = ConsumeType<Identifier>($"unexpected token: {tokens.First()} expected: ;");
-			return expr is null ? new Declarator(id.Name) : new Declarator(expr, id.Name);
+
+			if(Match(Keyword.Var))
+			{
+				Identifier varId = ConsumeType<Identifier>(expected: "identifier after var");
+				return new Declarator(varId.Name);
+			}
+
+
+			Expression expr = NextExpression();
+			if(Match(DelimeterToken.SemiC) || Tokens.Count == 0) return expr;
+			Identifier id = ConsumeType<Identifier>(expected: "; to close statement");
+			return new Declarator(expr, id.Name);
 		}
 	}
 }
