@@ -5,26 +5,18 @@ using System.Text;
 using System.Threading.Tasks;
 
 namespace Outlet.Tokens {
-	public class Delimeter : Token {
+	public class DelimeterToken : Symbol {
 
-		public static readonly Delimeter FuncParen = new Delimeter("(");
-		public static readonly Delimeter LeftParen = new Delimeter("(");
-		public static readonly Delimeter RightParen = new Delimeter(")");
-		public static readonly Delimeter IndexBrace = new Delimeter("[");
-		public static readonly Delimeter LeftBrace = new Delimeter("[");
-		public static readonly Delimeter RightBrace = new Delimeter("]");
-		public static readonly Delimeter RightCurly = new Delimeter("}");
-		public static readonly Delimeter LeftCurly = new Delimeter("{");
-		public static readonly Delimeter Comma = new Delimeter(",");
-		public static readonly Delimeter Colon = new Delimeter(":");
-		public static readonly Delimeter SemiC = new Delimeter(";");
+		private DelimeterToken(string name) : base(name) { }
 
-		public string Name;
-
-		private Delimeter(string name) {
-			Name = name;
-		}
-
-		public override string ToString() => Name;
+		public static readonly DelimeterToken LeftParen = new DelimeterToken("(");
+		public static readonly DelimeterToken RightParen = new DelimeterToken(")");
+		public static readonly DelimeterToken LeftBrace = new DelimeterToken("[");
+		public static readonly DelimeterToken RightBrace = new DelimeterToken("]");
+		public static readonly DelimeterToken RightCurly = new DelimeterToken("}");
+		public static readonly DelimeterToken LeftCurly = new DelimeterToken("{");
+		public static readonly DelimeterToken Comma = new DelimeterToken(",");
+		public static readonly DelimeterToken Colon = new DelimeterToken(":");
+		public static readonly DelimeterToken SemiC = new DelimeterToken(";");
 	}
 }
