@@ -6,28 +6,38 @@ namespace Outlet.Types
 {
     public class MethodGroupType : Type
     {
-        public readonly List<(FunctionType type, uint Id)> Methods;
+        // Hidden class so that not all FunctionTypes need an overload ID
+        private class MethodWrapper : IOverloadable
+        {
+            public MethodWrapper(FunctionType type, uint id)
+            {
+                OverloadId = id;
+                MethodType = type;
+            }
+
+            public uint OverloadId { get; private set; }
+            public FunctionType MethodType { get; private set; }
+
+            public bool Valid(out uint level, params Type[] inputs) => MethodType.Valid(out level, inputs);
+        }
+
+        private Overload<MethodWrapper> Methods { get; set; }
 
         public MethodGroupType(params (FunctionType type, uint id)[] functions)
         {
-            Methods = functions.ToList();
+            Methods = new Overload<MethodWrapper>(functions.Select(method => new MethodWrapper(method.type, method.id)).ToArray());
         }
+
+        public void AddMethod(FunctionType type, uint id) => Methods.Add(new MethodWrapper(type, id));
 
         public (FunctionType? type, uint? id) FindBestMatch(params Type[] inputs)
         {
-            (FunctionType? best, uint? bestLevel, uint? bestId) = (default, null, null);
-            foreach ((FunctionType overload, uint id) in Methods)
+            var method = Methods.FindBestMatch(inputs);
+            if(method != null)
             {
-                bool valid = overload.Valid(out uint level, inputs);
-                if (!valid) continue;
-                // TODO this logic may allow -1 to be set again
-                if (bestLevel == null || level < bestLevel)
-                {
-                    (best, bestLevel, bestId) = (overload, level, id);
-                }
+                return (method.MethodType, method.OverloadId);
             }
-
-            return (best, bestId);
+            return (null, null);
         }
 
         //public override bool Equals(Operand b) => ReferenceEquals(this, b);
